requirements.txt
beautifulsoup4>=4.9.3
requests>=2.22.0
schedule>=1.1.0
lxml>=4.6.3
python-dotenv>=0.19.0
discord.py>=1.7.3